requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cfgv==3.4.0
click==8.2.1
datetime==5.5
distlib==0.3.9
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
identify==2.6.12
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.16.0
mypy_extensions==1.1.0
nodeenv==1.9.1
pathspec==0.12.1
passlib==1.7.4
bcrypt==3.2.2
platformdirs==4.3.8
pre_commit==4.2.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyright==1.1.401
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
ruff==0.11.12
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1

